Move convert_gaze_detections_to_sv_detections_and_angles away from steps utils
